Match Task_00_8002988

+++ include/game/shared/rings_manager.h
@@ -35,7 +35,7 @@
 
 extern const u8 *const gSpritePosData_rings[NUM_LEVEL_IDS];
 
-void AddRings(s32 count);
+void AddRings(u16 count);
 void CreateStageRingsManager(void);
 void CreateCollectRingEffect(
     s32 param0, s32 param1); // NOTE: this *should* take s16 parameters, see comment inside CreateCollectRingEffect for more details!
